+++ src/main/java/interface_adapter/execute_buy/ExecuteBuyPresenter.java
@@ -23,13 +23,8 @@ public class ExecuteBuyPresenter implements ExecuteBuyOutputBoundary {
     }
 
     @Override
-    public void prepareStockNotFoundView() {
-
-    }
-
-    @Override
     public void prepareValidationErrorView() {
-
+        viewModel.firePropertyChanged("validationError");
     }