+++ src/main/java/use_case/execute_buy/ExecuteBuyOutputData.java
@@ -5,4 +5,6 @@ import entity.Portfolio;
 public record ExecuteBuyOutputData(
         double newBalance,
         Portfolio newPortfolio
-) {}
+) {
+}
+